Calculate Ieqtls

Calculate your own Cell Specific eQTLs.

This service calculates the gene-variant association for any given pair of gene and variant, which may or may not be significant.

This requires as input a GENCODE ID, GTEx variant ID, and tissue site detail ID.

By default, the calculation is based on the latest GTEx release.

GTEx Portal API documentation.

Usage

calculate_ieqtls(
  cellType,
  tissueSiteDetailId,
  gencodeId,
  variantId,
  datasetId = "gtex_v8"
)

Arguments

cellType

String. "Adipocytes", "Epithelial_cells", "Hepatocytes", "Keratinocytes", "Myocytes", "Neurons", "Neutrophils".

tissueSiteDetailId

String. The ID of the tissue of interest. Can be a GTEx specific ID (e.g. "Whole_Blood"; use get_tissue_site_detail() to see valid values) or an Ontology ID.

gencodeId

String. A Versioned GENCODE ID of a gene, e.g. "ENSG00000065613.9".

variantId

String. A gtex variant ID.

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".

Value

A tibble.

See also

Examples

# \dontrun{
# perform request
calculate_ieqtls(cellType = "Adipocytes",
                 tissueSiteDetailId = "Adipose_Subcutaneous",
                 gencodeId = "ENSG00000203782.5",
                 variantId = "chr1_1099341_T_C_b38")
#> # A tibble: 1 × 9
#>   cellType  data  datasetId enrichmentScores gencodeId genotypes regressionCoord
#>   <chr>     <lis> <chr>     <list>           <chr>     <list>    <list>         
#> 1 Adipocyt… <dbl> gtex_v8   <dbl [581]>      ENSG0000… <int>     <named list>   
#> # ℹ 2 more variables: tissueSiteDetailId <chr>, variantId <chr>
# }