Get Annotation

This service returns the list of annotations and allowed values by which a particular dataset can be subsetted. Results may be filtered by dataset.

GTEx Portal API documentation

Usage

get_annotation(datasetId = "gtex_v8", page = 0, itemsPerPage = 250)

Arguments

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A tibble

Details

Note: the output for this function appears to be incomplete currently.

See also

Examples

# \dontrun{
 get_annotation()
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 1
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 2
#> # A tibble: 2 × 5
#>   entity  values    type  annotation datasetId
#>   <chr>   <list>    <chr> <chr>      <chr>    
#> 1 subject <chr [2]> enum  sex        gtex_v8  
#> 2 subject <chr [6]> enum  ageBracket gtex_v8  
# }