Get Clustered Median Exon Expression

Find median transcript expression data along with hierarchical clusters.

Returns median normalized transcript expression in tissues of all known transcripts of a given gene along with the hierarchical clustering results of tissues and transcripts, based on exon expression, in Newick format.

The hierarchical clustering is performed by calculating Euclidean distances and using the average linkage method.

This endpoint is not paginated.

By default, this endpoint queries the latest GTEx release.

GTEx Portal API documentation

Usage

get_clustered_median_exon_expression(
  gencodeIds,
  datasetId = "gtex_v8",
  tissueSiteDetailIds = NULL
)

Arguments

gencodeIds

A character vector of Versioned GENCODE IDs, e.g. c("ENSG00000132693.12", "ENSG00000203782.5").

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".

tissueSiteDetailIds

Character vector of IDs for tissues of interest. Can be GTEx specific IDs (e.g. "Whole_Blood"; use available_tissueSiteDetailIds() to see valid values) or Ontology IDs.

Value

A tibble, with clustering data stored as an attribute, "clusters".

See also

Examples

# \dontrun{
get_clustered_median_exon_expression(c("ENSG00000203782.5",
                                       "ENSG00000132693.12"))
#> ℹ Retrieve clustering data with `attrbutes(<df>)$clusters`
#> # A tibble: 216 × 8
#>     median exonId   tissueSiteDetailId ontologyId datasetId gencodeId geneSymbol
#>      <dbl> <chr>    <chr>              <chr>      <chr>     <chr>     <chr>     
#>  1  0.789  ENSG000… Adipose_Subcutane… UBERON:00… gtex_v8   ENSG0000… CRP       
#>  2 16.8    ENSG000… Adipose_Subcutane… UBERON:00… gtex_v8   ENSG0000… CRP       
#>  3  0.0921 ENSG000… Adipose_Visceral_… UBERON:00… gtex_v8   ENSG0000… CRP       
#>  4 12      ENSG000… Adipose_Visceral_… UBERON:00… gtex_v8   ENSG0000… CRP       
#>  5  1      ENSG000… Adrenal_Gland      UBERON:00… gtex_v8   ENSG0000… CRP       
#>  6 22.7    ENSG000… Adrenal_Gland      UBERON:00… gtex_v8   ENSG0000… CRP       
#>  7  0      ENSG000… Artery_Aorta       UBERON:00… gtex_v8   ENSG0000… CRP       
#>  8  8      ENSG000… Artery_Aorta       UBERON:00… gtex_v8   ENSG0000… CRP       
#>  9  0      ENSG000… Artery_Coronary    UBERON:00… gtex_v8   ENSG0000… CRP       
#> 10 14      ENSG000… Artery_Coronary    UBERON:00… gtex_v8   ENSG0000… CRP       
#> # ℹ 206 more rows
#> # ℹ 1 more variable: unit <chr>

# clustering data is stored as an attribute "clusters"
attributes(result)$cluster
#> Error in eval(expr, envir, enclos): object 'result' not found
# }