Get Eqtl Genes

Retrieve eGenes (eQTL Genes).

This service returns eGenes (eQTL Genes) from the specified dataset.

eGenes are genes that have at least one significant cis-eQTL acting upon them.

Results may be filtered by tissue. By default, the service queries the latest GTEx release.

For each eGene, the results include the allelic fold change (log2AllelicFoldChange), p-value (pValue), p-value threshold (pValueThreshold), empirical p-value (empiricalPValue), and q-value (qValue).

The log2AllelicFoldChange is the allelic fold change (in log2 scale) of the most significant eQTL.

The pValue is the nominal p-value of the most significant eQTL.

The pValueThreshold is the p-value threshold used to determine whether a cis-eQTL for this gene is significant. For more details see https://gtexportal.org/home/documentationPage#staticTextAnalysisMethods.

The empiricalPValue is the beta distribution-adjusted empirical p-value from FastQTL.

The qValues were calculated based on the empirical p-values. A false discovery rate (FDR) threshold of <= 0.05 was applied to identify genes with a significant eQTL.

GTEx Portal API documentation.

Usage

get_eqtl_genes(
  tissueSiteDetailIds,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)

Arguments

tissueSiteDetailIds

Character vector of IDs for tissues of interest. Can be GTEx specific IDs (e.g. "Whole_Blood"; use get_tissue_site_detail() to see valid values) or Ontology IDs.

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A tibble.

See also

Examples

# \dontrun{
get_eqtl_genes(c("Whole_Blood", "Artery_Aorta"))
#> Warning: ! Total number of items (24853) exceeds maximum page size (250).
#> ℹ Try increasing `itemsPerPage`.
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 100
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 24853
#> # A tibble: 250 × 10
#>    tissueSiteDetailId ontologyId  datasetId empiricalPValue gencodeId geneSymbol
#>    <chr>              <chr>       <chr>               <dbl> <chr>     <chr>     
#>  1 Whole_Blood        UBERON:001… gtex_v8          1.05e- 9 ENSG0000… WASH7P    
#>  2 Whole_Blood        UBERON:001… gtex_v8          1.06e-25 ENSG0000… RP11-34P1…
#>  3 Whole_Blood        UBERON:001… gtex_v8          6.31e- 2 ENSG0000… CICP27    
#>  4 Whole_Blood        UBERON:001… gtex_v8          8.71e- 9 ENSG0000… RP11-34P1…
#>  5 Whole_Blood        UBERON:001… gtex_v8          6.01e-20 ENSG0000… RP11-34P1…
#>  6 Whole_Blood        UBERON:001… gtex_v8          6.96e- 9 ENSG0000… RP11-34P1…
#>  7 Whole_Blood        UBERON:001… gtex_v8          3.10e- 4 ENSG0000… RP11-34P1…
#>  8 Whole_Blood        UBERON:001… gtex_v8          1.92e- 3 ENSG0000… ABC7-4304…
#>  9 Whole_Blood        UBERON:001… gtex_v8          1.58e- 3 ENSG0000… RP11-34P1…
#> 10 Whole_Blood        UBERON:001… gtex_v8          7.82e- 2 ENSG0000… AP006222.2
#> # ℹ 240 more rows
#> # ℹ 4 more variables: log2AllelicFoldChange <dbl>, pValue <dbl>,
#> #   pValueThreshold <dbl>, qValue <dbl>
# }