Get Exons

This service returns exons from all known transcripts of the given gene.

A versioned GENCODE ID is required to ensure that all exons are from a single gene.

A dataset ID or both GENCODE version and genome build must be provided.

Although annotated exons are not dataset dependent, specifying a dataset here is equivalent to specifying the GENCODE version and genome build used by that dataset.

GTEx Portal API documentation

Usage

get_exons(
  gencodeId,
  gencodeVersion = NULL,
  genomeBuild = NULL,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)

Arguments

gencodeId

String. A Versioned GENCODE ID of a gene, e.g. "ENSG00000065613.9".

gencodeVersion

String (default = "v26"). GENCODE annotation release. Either "v26" or "v19".

genomeBuild

String. Options: "GRCh38/hg38", "GRCh37/hg19". Default = "GRCh38/hg38".

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A tibble.

See also

Examples

# \dontrun{
get_exons(gencodeId = "ENSG00000203782.5")
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 1
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 2
#> # A tibble: 2 × 13
#>   featureType       end genomeBuild     start exonId source chromosome gencodeId
#>   <chr>           <int> <chr>           <int> <chr>  <chr>  <chr>      <chr>    
#> 1 exon        153259733 GRCh38/hg38 153259700 ENSE0… HAVANA chr1       ENSG0000…
#> 2 exon        153262122 GRCh38/hg38 153260927 ENSE0… HAVANA chr1       ENSG0000…
#> # ℹ 5 more variables: transcriptId <chr>, geneSymbol <chr>,
#> #   gencodeVersion <chr>, strand <chr>, exonNumber <chr>
# }