Get Gene Search

Find genes that are partial or complete match of a gene_id

Usage

get_gene_search(
  geneId,
  gencodeVersion = "v26",
  genomeBuild = "GRCh38/hg38",
  page = 0,
  itemsPerPage = 250
)

Arguments

geneId

String.

gencodeVersion

String (default = "v26"). GENCODE annotation release. Either "v26" or "v19".

genomeBuild

String. Options: "GRCh38/hg38", "GRCh37/hg19", Default = "GRCh38/hg38".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A tibble.

Details

gene_id could be a gene symbol, a gencode ID, or an Ensemble ID

Gencode Version and Genome Build must be specified

See also

Examples

get_gene_search("CRP")
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 1
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 2
#> # A tibble: 2 × 15
#>   chromosome dataSource description      end gencodeId gencodeVersion geneStatus
#>   <chr>      <chr>      <chr>          <int> <chr>     <chr>          <chr>     
#> 1 chr1       HAVANA     C-reactive p… 1.60e8 ENSG0000… v26            ""        
#> 2 chr1       HAVANA     C-reactive p… 1.60e8 ENSG0000… v26            ""        
#> # ℹ 8 more variables: geneSymbol <chr>, geneSymbolUpper <chr>, geneType <chr>,
#> #   genomeBuild <chr>, start <int>, strand <chr>, tss <int>, entrezGeneId <chr>