Get Linkage Disequilibrium Data

Find linkage disequilibrium (LD) data for a given gene.

This endpoint returns linkage disequilibrium data for the cis-eQTLs found associated with the provided gene in a specified dataset.

Results are queried by gencode ID. By default, the service queries the latest GTEx release.

Specify a dataset ID to fetch results from a different dataset.

GTEx Portal API documentation

Usage

get_linkage_disequilibrium_data(
  gencodeIds,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)

Arguments

gencodeIds

A character vector of Versioned GENCODE IDs, e.g. c("ENSG00000132693.12", "ENSG00000203782.5").

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A tibble.

See also

Examples

get_linkage_disequilibrium_data(gencodeIds = c("ENSG00000132693.12",
                                                  "ENSG00000203782.5"))
#> # A tibble: 250 × 3
#>    variantId_1            variantId_2            ld                
#>    <chr>                  <chr>                  <chr>             
#>  1 chr1_152442732_A_G_b38 chr1_152442997_C_T_b38 0.9967540000000001
#>  2 chr1_152442732_A_G_b38 chr1_152444300_T_C_b38 1.0               
#>  3 chr1_152442732_A_G_b38 chr1_152445670_T_C_b38 1.0               
#>  4 chr1_152442732_A_G_b38 chr1_152445737_A_C_b38 1.0               
#>  5 chr1_152442732_A_G_b38 chr1_152446586_C_T_b38 1.0               
#>  6 chr1_152442732_A_G_b38 chr1_152447965_A_C_b38 1.0               
#>  7 chr1_152442732_A_G_b38 chr1_152448570_T_C_b38 0.9967360000000001
#>  8 chr1_152442732_A_G_b38 chr1_152448978_G_T_b38 0.99351           
#>  9 chr1_152442732_A_G_b38 chr1_152449069_C_A_b38 0.99351           
#> 10 chr1_152442732_A_G_b38 chr1_152449124_T_A_b38 0.99351           
#> # ℹ 240 more rows