Get Multi Tissue Eqtls

Find multi-tissue eQTL Metasoft results.

This service returns multi-tissue eQTL Metasoft results for a given gene and variant in a specified dataset.

A Versioned GENCODE ID must be provided.

For each tissue, the results include: m-value (mValue), normalized effect size (nes), p-value (pValue), and standard error (se).

The m-value is the posterior probability that an eQTL effect exists in each tissue tested in the cross-tissue meta-analysis (Han and Eskin, PLoS Genetics 8(3): e1002555, 2012).

The normalized effect size is the slope of the linear regression of normalized expression data versus the three genotype categories using single-tissue eQTL analysis, representing eQTL effect size.

The p-value is from a t-test that compares observed NES from single-tissue eQTL analysis to a null NES of 0.

By default, the service queries the latest GTEx release. The retrieved data is split into pages with items_per_page entries per page

GTEx Portal API documentation

Usage

get_multi_tissue_eqtls(
  gencodeIds,
  variantId = NULL,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)

Arguments

gencodeIds

A character vector of Versioned GENCODE IDs, e.g. c("ENSG00000132693.12", "ENSG00000203782.5").

variantId

String. A gtex variant ID.

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot". Default = "gtex_v8".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A tibble.

See also

Examples

# search by gene
get_multi_tissue_eqtls(gencodeId = c("ENSG00000132693.12",
                                     "ENSG00000203782.5"))
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 1
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 132
#> # A tibble: 6,468 × 5
#>    gencodeId         datasetId   metaP variantId              tissues         
#>    <chr>             <chr>       <dbl> <chr>                  <named list>    
#>  1 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#>  2 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#>  3 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#>  4 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#>  5 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#>  6 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#>  7 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#>  8 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#>  9 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#> 10 ENSG00000203782.5 gtex_v8   0.00177 chr1_152346526_G_T_b38 <named list [4]>
#> # ℹ 6,458 more rows