Get News Item

Getting all the news items from the database that are current.

Usage

get_news_item(page = 0, itemsPerPage = 250)

Arguments

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

List.

Examples

get_news_item()
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 1
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 23
#> $data
#> $data[[1]]
#> $data[[1]]$title
#> [1] "New Top 50 Expressed Genes Visualization"
#> 
#> $data[[1]]$newsText
#> [1] "The GTEx Portal has a new Top 50 Expressed Genes visualization.  This visualization allows you to select your tissue of interest and display the top 50 expressed genes in a heat map showing the expression of those genes across all tissues.Mitochondrial genes are excluded from the list by default, but can be included. Genes and tissues are clustered using hierarchical clustering. One or more genes can be selected on the heatmap, which will display comparative expression box plots of that gene across all tissues. You can download the figure for use in your publication.<br/>You can view the top 50 expressed genes page here: <a href=\"https://gtexportal.org/home/topExpressedGenePage\">https://gtexportal.org/home/topExpressedGenePage</a>"
#> 
#> $data[[1]]$rank
#> [1] 240
#> 
#> $data[[1]]$dateCreated
#> [1] "2018-04-04"
#> 
#> $data[[1]]$release
#> [1] "news_item"
#> 
#> $data[[1]]$id
#> [1] 240
#> 
#> 
#> $data[[2]]
#> $data[[2]]$title
#> [1] "New Multi-Gene Query Visualization"
#> 
#> $data[[2]]$newsText
#> [1] "The GTEx Portal has a new Multi-Gene Query visualization.  This visualization allows you to enter a list of genes of interest and display the expression of those genes in a heat map that shows expression across all GTEx tissues. By default, the genes and tissues are clustered using hierarchical clustering.  One or more genes can be selected on the heatmap, which will display comparative expression box plots of that gene across all tissues. You can download the figure for use in your publication.<br/>You can view the Multi-Gene Query page here: <a href=\"https://gtexportal.org/home/multiGeneQueryPage\">https://gtexportal.org/home/multiGeneQueryPage</a>"
#> 
#> $data[[2]]$rank
#> [1] 250
#> 
#> $data[[2]]$dateCreated
#> [1] "2018-04-04"
#> 
#> $data[[2]]$release
#> [1] "news_item"
#> 
#> $data[[2]]$id
#> [1] 250
#> 
#> 
#> $data[[3]]
#> $data[[3]]$title
#> [1] "New eQTL Dashboard Visualization"
#> 
#> $data[[3]]$newsText
#> [1] "The GTEx Portal has a new eQTL Dashboard visualization.  This visualization allows you to view eQTLs for your selected gene-variant pairs across a selected set of tissues.<br/>You can view the eQTL Dashboard page here: <a href=\"https://gtexportal.org/home/eqtlDashboardPage\">https://gtexportal.org/home/eqtlDashboardPage</a>"
#> 
#> $data[[3]]$rank
#> [1] 260
#> 
#> $data[[3]]$dateCreated
#> [1] "2018-04-04"
#> 
#> $data[[3]]$release
#> [1] "news_item"
#> 
#> $data[[3]]$id
#> [1] 260
#> 
#> 
#> $data[[4]]
#> $data[[4]]$title
#> [1] "Video Tutorial for the Gene-eQTL Visualizer"
#> 
#> $data[[4]]$newsText
#> [1] "Watch our new video tutorial for the Gene-eQTL Visualizer. <a href=\"https://gtexportal.org/home/videos\">https://gtexportal.org/home/videos</a>"
#> 
#> $data[[4]]$rank
#> [1] 270
#> 
#> $data[[4]]$dateCreated
#> [1] "2018-04-05"
#> 
#> $data[[4]]$release
#> [1] "news_item"
#> 
#> $data[[4]]$id
#> [1] 270
#> 
#> 
#> $data[[5]]
#> $data[[5]]$title
#> [1] "Access GTEx Biospecimens Here"
#> 
#> $data[[5]]$newsText
#> [1] "We have updated our biospecimen search page.  New features include a shopping cart for streamlined sample requests: <a href=\"https://gtexportal.org/home/samplesPage\">https://gtexportal.org/home/samplesPage</a>"
#> 
#> $data[[5]]$rank
#> [1] 280
#> 
#> $data[[5]]$dateCreated
#> [1] "2018-05-10"
#> 
#> $data[[5]]$release
#> [1] "news_item"
#> 
#> $data[[5]]$id
#> [1] 280
#> 
#> 
#> $data[[6]]
#> $data[[6]]$title
#> [1] "Access GTEx Data Programmatically with the GTEx API"
#> 
#> $data[[6]]$newsText
#> [1] "You can now access GTEx public-access data programmatically using the GTEx API.  This REST API allows you to retrieve expression, eQTL, and annotation data by constructing simple URLs.  You can find documentation for our beta release of the API here: <a href=\"https://gtexportal.org/home/api-docs\">https://gtexportal.org/home/api-docs</a>"
#> 
#> $data[[6]]$rank
#> [1] 290
#> 
#> $data[[6]]$dateCreated
#> [1] "2018-09-20"
#> 
#> $data[[6]]$release
#> [1] "news_item"
#> 
#> $data[[6]]$id
#> [1] 290
#> 
#> 
#> $data[[7]]
#> $data[[7]]$title
#> [1] "Reuse GTEx Visualizations On Your Website"
#> 
#> $data[[7]]$newsText
#> [1] "You can now embed GTEx visualizations on your website.  GTEx visualizations are now open source with documentation on how you can embed them into your web pages.  The source code for the visualizations and documentation are available here: <a href=\"https://github.com/broadinstitute/gtex-viz\">https://github.com/broadinstitute/gtex-viz</a>."
#> 
#> $data[[7]]$rank
#> [1] 300
#> 
#> $data[[7]]$dateCreated
#> [1] "2018-09-20"
#> 
#> $data[[7]]$release
#> [1] "news_item"
#> 
#> $data[[7]]$id
#> [1] 300
#> 
#> 
#> $data[[8]]
#> $data[[8]]$title
#> [1] "Use The GTEx Portal On Your Phone"
#> 
#> $data[[8]]$newsText
#> [1] "We have released the GTEx mobile website.  This mobile site optimizes the GTEx experience for your phone or tablet, making it easy to look up gene expression and eQTLs from a mobile device.  Visit <a href=\"https://gtexportal.org\">https://gtexportal.org</a> from your phone, and it will automatically forward to the mobile site.  Don’t worry, you can still get to the full site from the mobile site -- there is a link at the bottom."
#> 
#> $data[[8]]$rank
#> [1] 310
#> 
#> $data[[8]]$dateCreated
#> [1] "2018-09-20"
#> 
#> $data[[8]]$release
#> [1] "news_item"
#> 
#> $data[[8]]$id
#> [1] 310
#> 
#> 
#> $data[[9]]
#> $data[[9]]$title
#> [1] "GTEx Portal Visualizations Updated"
#> 
#> $data[[9]]$newsText
#> [1] "We have updated several visualizations in the GTEx Portal.  We have replaced our gene expression and single-tissue eQTL box plots with violin plots.  Our single-tissue eQTL violin plot interface now allows you to view multiple single-tissue eQTLs at once.<br/><br/>We have completely rewritten our Gene-eQTL Visualizer (GEV).  The central bubble-heatmap field now zooms to fit the viewport into the available screen size.  We have improved GEV performance, which makes the viewport selector and filtering by p-value and normalized effect size more responsive.  The linkage-disequilibrium display has been adjusted to be more easily interpreted."
#> 
#> $data[[9]]$rank
#> [1] 320
#> 
#> $data[[9]]$dateCreated
#> [1] "2018-12-06"
#> 
#> $data[[9]]$release
#> [1] "news_item"
#> 
#> $data[[9]]$id
#> [1] 320
#> 
#> 
#> $data[[10]]
#> $data[[10]]$title
#> [1] "Help Us Help You: New Feature Survey"
#> 
#> $data[[10]]$newsText
#> [1] "PLEASE TAKE OUR SURVEY to help us plan new features for the GTEx Portal: <a href=\"http://bit.ly/2UMEgbP\" target=\"_blank\">http://bit.ly/2UMEgbP</a>"
#> 
#> $data[[10]]$rank
#> [1] 331
#> 
#> $data[[10]]$dateCreated
#> [1] "2019-02-06"
#> 
#> $data[[10]]$release
#> [1] "news_item"
#> 
#> $data[[10]]$id
#> [1] 331
#> 
#> 
#> $data[[11]]
#> $data[[11]]$title
#> [1] "Retina expression and eQTLs from EyeGEx released"
#> 
#> $data[[11]]$newsText
#> [1] "Retina gene expression and eQTLs from the Eye Genotype Expression (EyeGEx) database (<a href=\"https://www.nature.com/articles/s41588-019-0351-9\">https://www.nature.com/articles/s41588-019-0351-9</a>) are now available for download on the GTEx portal under External Datasets. Postmortem retina samples were taken from non-diseased and diseased eyes at different stages of age-related macular degeneration (AMD)."
#> 
#> $data[[11]]$rank
#> [1] 335
#> 
#> $data[[11]]$dateCreated
#> [1] "2019-02-11"
#> 
#> $data[[11]]$release
#> [1] "news_item"
#> 
#> $data[[11]]$id
#> [1] 335
#> 
#> 
#> $data[[12]]
#> $data[[12]]$title
#> [1] "New Histology Image Viewer"
#> 
#> $data[[12]]$newsText
#> [1] "Check out our new revised histology image viewer.  The streamlined interface makes it easy to search by sample identifiers.  The advanced search feature allows you to easily search for histology images by donor phenotype or tissue type, but also hides away when you don't need it.<br/>Our new multi-image mode allows you to view multiple histology images side-by-side.<br/><a href='https://gtexportal.org/home/histologyPage'>https://gtexportal.org/home/histologyPage</a>"
#> 
#> $data[[12]]$rank
#> [1] 340
#> 
#> $data[[12]]$dateCreated
#> [1] "2019-03-07"
#> 
#> $data[[12]]$release
#> [1] "news_item"
#> 
#> $data[[12]]$id
#> [1] 340
#> 
#> 
#> $data[[13]]
#> $data[[13]]$title
#> [1] "GTEx V8 data release"
#> 
#> $data[[13]]$newsText
#> [1] "The GTEx V8 protected data is available on dbGaP. Access to the raw sequence data, formerly available on the SRA, is now provided through the AnVIL platform (all other protected data are available in this repository as well). For access instructions please see <a href=\"https://gtexportal.org/home/v8ReleasePage\">link</a>.<br> GTEx V8 public data, including QTLs and expression tables, are currently being integrated into the GTEx Portal. These data will be released on the GTEx Portal in August, 2019."
#> 
#> $data[[13]]$rank
#> [1] 350
#> 
#> $data[[13]]$dateCreated
#> [1] "2019-07-24"
#> 
#> $data[[13]]$release
#> [1] "news_item"
#> 
#> $data[[13]]$id
#> [1] 350
#> 
#> 
#> $data[[14]]
#> $data[[14]]$title
#> [1] "GTEx Portal V8 Release"
#> 
#> $data[[14]]$newsText
#> [1] "The GTEx V8 non-protected data is now available on the GTEx Portal.  Release V8 includes 17,382 RNA-Seq samples from 948 donors, representing an increase of 49% and 33% relative to V7, respectively.<br/>Release V8 includes a new quantitative trait loci data type: splicing QTLs (sQTL).  We have replaced the Gene eQTL Visualizer with a new <a href=\"https://gtexportal.org/home/locusBrowserPage/RSPRY1\" target=\"_blank\">Locus Browser</a> that displays both eQTLs and sQTLs.<br/> The tissue previously named \"Cells - Transformed fibroblasts\" has been corrected and is now named \"Cells - Cultured fibroblasts\".<br/>The visualizations on the GTEx Portal have been updated to the Release V8 data. The Release V8 public data are available for download <a href=\"https://gtexportal.org/home/datasets\" target=\"_blank\">here</a>."
#> 
#> $data[[14]]$rank
#> [1] 360
#> 
#> $data[[14]]$dateCreated
#> [1] "2019-08-26"
#> 
#> $data[[14]]$release
#> [1] "news_item"
#> 
#> $data[[14]]$id
#> [1] 360
#> 
#> 
#> $data[[15]]
#> $data[[15]]$title
#> [1] "RNA-seq Coverage Tracks"
#> 
#> $data[[15]]$newsText
#> [1] "We have added tissue-specific RNA-seq coverage tracks to the GTEx IGV Browser.  The revised track menu allows you to seleect the coverage tracks.<br/><br/>Also, we have upgraded the version of IGV and fixed bugs in the GTEx IGV Browser track menu. <a href='https://gtexportal.org/home/browseEqtls?location=chr1%3A1-500,000'>GTEx IGV Browser</a>"
#> 
#> $data[[15]]$rank
#> [1] 380
#> 
#> $data[[15]]$dateCreated
#> [1] "2020-01-06"
#> 
#> $data[[15]]$release
#> [1] "news_item"
#> 
#> $data[[15]]$id
#> [1] 380
#> 
#> 
#> $data[[16]]
#> $data[[16]]$title
#> [1] "New Functional Annotation Track in Locus Browser"
#> 
#> $data[[16]]$newsText
#> [1] "We have added a new functional annotation track in the GTEx Locus Browser.  This track displays the functional annotation for the variants in view on the Locus Browser.<br/><br/>Functional annotation of variants within transcript regions was performed using Ensembl's Variant Effect Predictor (VEP) and Loss-Of-Function Transcript Effect Estimator (LOFTEE) (VEP v85, GENCODE v26) implemented in Hail v0.1.<br/></br><a href='https://gtexportal.org/home/locusBrowserPage/RSPRY1'>GTEx Locus Browser</a>"
#> 
#> $data[[16]]$rank
#> [1] 400
#> 
#> $data[[16]]$dateCreated
#> [1] "2020-01-06"
#> 
#> $data[[16]]$release
#> [1] "news_item"
#> 
#> $data[[16]]$id
#> [1] 400
#> 
#> 
#> $data[[17]]
#> $data[[17]]$title
#> [1] "NHGRI AnVIL Cloud Platform Now Supports Free Export of GTEx Data"
#> 
#> $data[[17]]$newsText
#> [1] "One of the most widely-used resources for studying the relationship between genetic variation and gene expression is the <a href='https://www.genome.gov/Funded-Programs-Projects/Genotype-Tissue-Expression-Project'>Genotype-Tissue Expression (GTEx) project</a>.  Established by the NIH Common Fund in 2010, the recent <a href='https://science.sciencemag.org/content/369/6509/1318'>GTEx V8 dataset</a> represents the largest atlas of human gene expression and corresponding trait loci to date (dbGaP accession: <a href='https://www.ncbi.nlm.nih.gov/projects/gap/cgi-bin/study.cgi?study_id=phs000424'>phs000424.v8.p2</a>).  This dataset contains genotype data from 838 postmortem donors and 17,382 RNA-seq samples across 54 tissue sites and 2 cell lines.  The <a href='https://gtexportal.org'>GTEx Portal</a> provides uniformly processed gene expression data, a QTL Browser, and a mechanism by which to request available biospecimens to allow researchers to study the impact of genetic variation on complex traits and diseases.<br><br>Controlled access to raw and protected data that may identify donors is provided through the <a href='https://anvilproject.org'>AnVIL Project</a>.  Established in 2018, the NHGRI Genomic Data Science Analysis, Visualization, and Informatics Lab-space (AnVIL) provides a cloud-based environment that colocates an extensive collection of high-value datasets with commonly used bioinformatics tools in a secure computing environment.<br><br>In support of the evolving nature of the NHGRI mission, we are pleased to announce that researchers are now able to download controlled access GTEx V8 to local compute infrastructure without incurring egress fees. Researchers will still need to apply for approval from dbGaP to access these data and maintain data security on their institutional clusters, but this new capability will save researchers substantial expenses, especially for researchers that wish to integrate GTEx with their own clinical research data. Download instructions can be found at <a href='https://anvilproject.org/learn/reference/gtex-v8-free-egress-instructions'>https://anvilproject.org/learn/reference/gtex-v8-free-egress-instructions</a>.<br><br>We foresee that in the long term, more users will choose to perform the analysis of GTEx and other large datasets directly within AnVIL’s cloud environment.  AnVIL offers an elastic, shared computing resource, with active threat detection and monitoring, that provides an increasingly attractive alternative to redundantly downloaded data amongst siloed compute infrastructure.<br><br>AnVIL is built on a set of established components that have been used in a number of flagship scientific projects. The <a href='https://anvil.terra.bio/'>Terra platform</a> provides a compute environment with secure data and analysis sharing capabilities. <a href='https://dockstore.org/'>Dockstore</a> provides standards based sharing of containerized tools and workflows. <a href='https://www.bioconductor.org/'>R/Bioconductor</a> and <a href='https://galaxyproject.org/'>Galaxy</a> provide environments for users at different skill levels to construct and execute analyses. The <a href='https://gen3.theanvil.io/'>Gen3 data commons</a> framework provides data and metadata ingest, querying, and organization. Taken together, the AnVIL cloud platform allows researchers to build novel cohorts out of GTEx and other major NHGRI datasets such as <a href='https://www.genome.gov/Funded-Programs-Projects/NHGRI-Genome-Sequencing-Program/Centers-for-Common-Disease-Genomics'>CCDG</a>, <a href='https://www.genome.gov/Funded-Programs-Projects/NHGRI-Genome-Sequencing-Program/Centers-for-Mendelian-Genomics-CMG'>CMG</a>, and <a href='https://www.genome.gov/Funded-Programs-Projects/Electronic-Medical-Records-and-Genomics-Network-eMERGE'>eMERGE</a> and compute on them in place with scalable and easy to use tools."
#> 
#> $data[[17]]$rank
#> [1] 410
#> 
#> $data[[17]]$dateCreated
#> [1] "2020-11-20"
#> 
#> $data[[17]]$release
#> [1] "news_item"
#> 
#> $data[[17]]$id
#> [1] 410
#> 
#> 
#> $data[[18]]
#> $data[[18]]$title
#> [1] "Help Us Help You: New Feature Survey"
#> 
#> $data[[18]]$newsText
#> [1] "Please take our very brief user survey to help us plan new features for the GTEx Portal: <a href=\"http://bit.ly/2021gtexsurvey\" target=\"_blank\">http://bit.ly/2021gtexsurvey</a>"
#> 
#> $data[[18]]$rank
#> [1] 420
#> 
#> $data[[18]]$dateCreated
#> [1] "2021-04-22"
#> 
#> $data[[18]]$release
#> [1] "news_item"
#> 
#> $data[[18]]$id
#> [1] 420
#> 
#> 
#> $data[[19]]
#> $data[[19]]$title
#> [1] "snRNA-Seq Data released"
#> 
#> $data[[19]]$newsText
#> [1] "We have released snRNA-Seq data using 8 tissues and 16 GTEx donors (25 frozen tissue samples total).  This data is described on the new Single Cell Overview page and in <a href=\"https://www.biorxiv.org/content/10.1101/2021.07.19.452954v1\", target=\"_blank\">Eraslan et al., BioRxiv, 2021 <i class=\"fas fa-external-link-alt\"></i></a>.<br><br>Gene expression matrices are available for download from the <a href=\"https://gtexportal.org/home/datasets\">datasets page</a>.  Raw sequence data are available on <a href=\"https://anvil.terra.bio\" target=\"_blank\">AnVIL <i class=\"fas fa-external-link-alt\"></i></a>.<br><br>Single cell expression data for a single gene is now reported on the <a href=\"https://gtexportal.org/home/gene/PTEN\">gene page</a>.  The new <a href=\"https://gtexportal.org/home/multiGeneSingleCellQueryPage\">Multi-Gene Single Cell Viewer</a> displays cell expression data for a list of genes."
#> 
#> $data[[19]]$rank
#> [1] 430
#> 
#> $data[[19]]$dateCreated
#> [1] "2021-07-20"
#> 
#> $data[[19]]$release
#> [1] "news_item"
#> 
#> $data[[19]]$id
#> [1] 430
#> 
#> 
#> $data[[20]]
#> $data[[20]]$title
#> [1] "GTEx implements GA4GH’s RNAget API"
#> 
#> $data[[20]]$newsText
#> [1] "The GTEx Portal RNAget API provides the ability to retrieve GTEx RNA-seq data using the <a href='https://www.ga4gh.org/' target=_blank>GA4GH</a> standard <a href='https://github.com/ga4gh-rnaseq/schema' target=_blank>RNAget API</a>.  This API currently supports the retrieval of data for the GTEx V8 release.  You can find documentation for our implementation of RNAget <a href='https://gtexportal.org/home/apiPage'>here</a>. "
#> 
#> $data[[20]]$rank
#> [1] 440
#> 
#> $data[[20]]$dateCreated
#> [1] "2022-03-23"
#> 
#> $data[[20]]$release
#> [1] "news_item"
#> 
#> $data[[20]]$id
#> [1] 440
#> 
#> 
#> $data[[21]]
#> $data[[21]]$title
#> [1] "Cross-tissue Cartography"
#> 
#> $data[[21]]$newsText
#> [1] "Knowing where genes are active in the body is key to studying the full range of human diseases. Gökcen Eraslan, Eugene Drokhlyansky, Ayellet Segrè, François Aguet, Orit Rozenblatt-Rosen, Kristin Ardlie, and Aviv Regev developed a single-nucleus method that can profile myriad cell types, including cells from frozen tissue. With data from more than 200,000 cells across multiple tissues, the atlas revealed unexpected new cellular functions and linked some cells to specific diseases for the first time. Appearing in <a href='http://dx.doi.org/10.1126/science.abl4290' target=_blank>Science</a>, the paper is one of four from the Human Cell Atlas that describe complementary cross-tissue cell atlases. The data from this paper is available on the GTEx Portal <a href='https://gtexportal.org/home/singleCellOverviewPage'>here</a>. Read more in a <a href='https://www.broadinstitute.org/node/1132261' target=_blank>Broad story</a>."
#> 
#> $data[[21]]$rank
#> [1] 450
#> 
#> $data[[21]]$dateCreated
#> [1] "2022-05-12"
#> 
#> $data[[21]]$release
#> [1] "news_item"
#> 
#> $data[[21]]$id
#> [1] 450
#> 
#> 
#> $data[[22]]
#> $data[[22]]$title
#> [1] "GTEx Samples Mapped to HuBMAP Common Coordinate Framework"
#> 
#> $data[[22]]$newsText
#> [1] "We have spatially mapped GTEx samples from 16 tissue sites to the <a href='https://hubmapconsortium.org/' target='_blank'>HuBMAP <i class='fas fa-external-link-alt'></i></a> Common Coordinate Framework. We have embedded the HuBMAP <a href='https://gtexportal.org/home/eui'>Exploration User Inteface (EUI)</a> in the GTEx Portal, allowing users to view GTEx and HuBMAP samples in a 3D image of the human body. The mapped GTEx samples are also available on the <a href='https://portal.hubmapconsortium.org/' target='_blank'>HuBMAP Portal <i class='fas fa-external-link-alt'></i></a>.<br><br> Using the EUI, you can identify HuBMAP samples which are adjacent to GTEx sampling locations, and view data from those studies. We have a tutorial video that demonstrates how to use the EUI <a href='https://gtexportal.org/home/videos'>here</a>."
#> 
#> $data[[22]]$rank
#> [1] 460
#> 
#> $data[[22]]$dateCreated
#> [1] "2022-07-21"
#> 
#> $data[[22]]$release
#> [1] "news_item"
#> 
#> $data[[22]]$id
#> [1] 460
#> 
#> 
#> $data[[23]]
#> $data[[23]]$title
#> [1] "GTEx API V2 Released, GTEX API V1 Deprecated"
#> 
#> $data[[23]]$newsText
#> [1] "We have released the GTEx API V2. This version of our API provides improved query parameter validation, improved support for querying by tissue ontology identifiers, support for paginated queries, and improved documentation. The documentation for the GTEx API V2 can be found here: <a href='https://gtexportal.org/api/v2/redoc#tag/GTEx-Portal-API-Info' target='_blank'>here</a>. Our upcoming V10 data release will only be supported by the GTEx API V2. <br><br>We have deprecated our GTEx API V1 and it will no longer be operational on October 1, 2023.  Please update your code to use the GTEx API V2. The documentation for the GTEx API V1 can be found <a href='https://www.gtexportal.org/home/api-docs/#' target='_blank'>here</a>. Our upcoming V10 data release will not be available using the GTEx API V1."
#> 
#> $data[[23]]$rank
#> [1] 500
#> 
#> $data[[23]]$dateCreated
#> [1] "2023-02-23"
#> 
#> $data[[23]]$release
#> [1] "news_item"
#> 
#> $data[[23]]$id
#> [1] 500
#> 
#> 
#> 
#> $paging_info
#> $paging_info$numberOfPages
#> [1] 1
#> 
#> $paging_info$page
#> [1] 0
#> 
#> $paging_info$maxItemsPerPage
#> [1] 250
#> 
#> $paging_info$totalNumberOfItems
#> [1] 23
#> 
#>