Get Significant Single Tissue eQTLs By Location

Find significant single tissue eQTLs using Chromosomal Locations.

This service returns precomputed significant single tissue eQTLs.

Results may be filtered by tissue, and/or dataset.

By default, the service queries the latest GTEx release. Since this endpoint is used to support a third party program on the portal, the return structure is different from other endpoints and is not paginated.

GTEx Portal API documentation

Usage

get_significant_single_tissue_eqtls_by_location(
  tissueSiteDetailId,
  start,
  end,
  chromosome,
  datasetId = "gtex_v8"
)

Arguments

tissueSiteDetailId

String. The ID of the tissue of interest. Can be a GTEx specific ID (e.g. "Whole_Blood"; use get_tissue_site_detail() to see valid values) or an Ontology ID.

start

Integer.

end

Integer.

chromosome

String. One of "chr1", "chr2", "chr3", "chr4", "chr5", "chr6", "chr7", "chr8", "chr9", "chr10", "chr11", "chr12", "chr13", "chr14", "chr15", "chr16", "chr17", "chr18", "chr19", "chr20", "chr21", "chr22", "chrM", "chrX", "chrY".

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".

Value

A tibble.

See also

Examples

# \dontrun{
get_significant_single_tissue_eqtls_by_location(
  tissueSiteDetailId = "Artery_Aorta",
  start = 10000,
  end = 250000,
  chromosome = "chr11")
#> # A tibble: 434 × 11
#>    chromosome datasetId gencodeId      geneSymbol geneSymbolUpper   nes   pValue
#>    <chr>      <chr>     <chr>          <chr>      <chr>           <dbl>    <dbl>
#>  1 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.381 1.11e- 6
#>  2 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.364 2.25e- 6
#>  3 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.417 3.10e- 8
#>  4 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.480 8.39e- 9
#>  5 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.462 6.36e- 9
#>  6 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.553 8.89e-10
#>  7 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.577 2.95e-12
#>  8 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.547 6.13e-10
#>  9 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.462 6.36e- 9
#> 10 chr11      gtex_v8   ENSG000001779… BET1L      BET1L           0.462 6.36e- 9
#> # ℹ 424 more rows
#> # ℹ 4 more variables: pos <int>, snpId <chr>, tissueSiteDetailId <chr>,
#> #   variantId <chr>
# }