Get Significant Single Tissue Isqtls

Retrieve Interaction sQTL Data. This service retrieves cell type interaction sQTLs (isQTLs), from a specified dataset. Results may be filtered by tissue By default, the service queries the latest GTEx release. The retrieved data is split into pages with items_per_page entries per page

Usage

get_significant_single_tissue_isqtls(
  gencodeId,
  variantId = NULL,
  tissueSiteDetailId = NULL,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)

Arguments

gencodeId

String. A Versioned GENCODE ID of a gene, e.g. "ENSG00000065613.9".

variantId

String. A gtex variant ID.

tissueSiteDetailId

String. The ID of the tissue of interest. Can be a GTEx specific ID (e.g. "Whole_Blood"; use available_tissueSiteDetailIds() to see valid values) or an Ontology ID.

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot". Default = "gtex_v8".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A Tibble

Details

GTEx Portal API documentation

See also

Examples

# \dontrun{
get_significant_single_tissue_isqtls(gencodeId = "ENSG00000132693.12")
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 0
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 0
#> # A tibble: 0 × 0
# }