Get Single Nucleux Gex

Retrieve Single Nucleus Gene Expression Data for a given Gene.

GTEx Portal API documentation

Usage

get_single_nucleus_gex(
  gencodeIds,
  datasetId = "gtex_snrnaseq_pilot",
  tissueSiteDetailIds = NULL,
  excludeDataArray = TRUE,
  page = 0,
  itemsPerPage = 250
)

Arguments

gencodeIds

A character vector of Versioned GENCODE IDs, e.g. c("ENSG00000132693.12", "ENSG00000203782.5").

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".

tissueSiteDetailIds

Character vector of IDs for tissues of interest. Can be GTEx specific IDs (e.g. "Whole_Blood"; use available_tissueSiteDetailIds() to see valid values) or Ontology IDs.

excludeDataArray

String. Options are TRUE or FALSE

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A Tibble

See also

Examples

# \dontrun{
# Search for one or more genes
get_single_nucleus_gex(gencodeIds = c("ENSG00000203782.5",
                                      "ENSG00000132693.12"))
#> # A tibble: 14 × 7
#>    tissueSiteDetailId  ontologyId datasetId gencodeId geneSymbol cellTypes unit 
#>    <chr>               <chr>      <chr>     <chr>     <chr>      <list>    <chr>
#>  1 Muscle_Skeletal     UBERON:00… gtex_snr… ENSG0000… LOR        <tibble>  log(…
#>  2 Breast_Mammary_Tis… UBERON:00… gtex_snr… ENSG0000… LOR        <tibble>  log(…
#>  3 Esophagus_Mucosa    UBERON:00… gtex_snr… ENSG0000… LOR        <tibble>  log(…
#>  4 Esophagus_Muscular… UBERON:00… gtex_snr… ENSG0000… LOR        <tibble>  log(…
#>  5 Lung                UBERON:00… gtex_snr… ENSG0000… LOR        <tibble>  log(…
#>  6 Prostate            UBERON:00… gtex_snr… ENSG0000… LOR        <tibble>  log(…
#>  7 Skin_Sun_Exposed_L… UBERON:00… gtex_snr… ENSG0000… LOR        <tibble>  log(…
#>  8 Muscle_Skeletal     UBERON:00… gtex_snr… ENSG0000… CRP        <tibble>  log(…
#>  9 Breast_Mammary_Tis… UBERON:00… gtex_snr… ENSG0000… CRP        <tibble>  log(…
#> 10 Esophagus_Mucosa    UBERON:00… gtex_snr… ENSG0000… CRP        <tibble>  log(…
#> 11 Esophagus_Muscular… UBERON:00… gtex_snr… ENSG0000… CRP        <tibble>  log(…
#> 12 Heart_Left_Ventric… UBERON:00… gtex_snr… ENSG0000… CRP        <tibble>  log(…
#> 13 Lung                UBERON:00… gtex_snr… ENSG0000… CRP        <tibble>  log(…
#> 14 Prostate            UBERON:00… gtex_snr… ENSG0000… CRP        <tibble>  log(…

# Returns a tibble with one row per tissue
response <- get_single_nucleus_gex(gencodeIds = "ENSG00000132693.12",
                                   excludeDataArray = FALSE,
                                   itemsPerPage = 2)

response
#> # A tibble: 2 × 7
#>   tissueSiteDetailId   ontologyId datasetId gencodeId geneSymbol cellTypes unit 
#>   <chr>                <chr>      <chr>     <chr>     <chr>      <list>    <chr>
#> 1 Muscle_Skeletal      UBERON:00… gtex_snr… ENSG0000… CRP        <tibble>  log(…
#> 2 Breast_Mammary_Tiss… UBERON:00… gtex_snr… ENSG0000… CRP        <tibble>  log(…

# Column "cellTypes" contains a tibble of data with one row for each cell type
# e.g. for Breast_Mammary_Tissue,
response$cellTypes[[2]]
#> # A tibble: 2 × 8
#>   cellType                  count meanWithZeros meanWithoutZeros medianWithZeros
#>   <chr>                     <int>         <dbl>            <dbl>           <dbl>
#> 1 Epithelial cell (luminal)     2      0.000902             2.19               0
#> 2 Endothelial cell (vascul…     1      0.00401              3.05               0
#> # ℹ 3 more variables: medianWithoutZeros <dbl>, numZeros <int>, data <list>

# Expression values are stored in "data" e.g. for Breast_Mammary_Tissue,
# Epithelial cell (luminal):
response$cellTypes[[2]]$data[[1]]
#> [1] 1.593315 2.782725
# }