Get sqtl Genes

Retrieve sGenes (sQTL Genes).

Usage

get_sqtl_genes(
  tissueSiteDetailId,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)

Arguments

tissueSiteDetailId

String. The ID of the tissue of interest. Can be a GTEx specific ID (e.g. "Whole_Blood"; use available_tissueSiteDetailIds() to see valid values) or an Ontology ID.

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A tibble.

Details

This service returns sGenes (sQTL Genes) from the specified dataset.

sGenes are genes that have at least one significant cis-sQTL acting upon them.

Results may be filtered by tissue. By default, the service queries the latest GTEx release.

For each sGene, the results include the allelic fold change (log2AllelicFoldChange), p-value (pValue), p-value threshold (pValueThreshold), empirical p-value (empiricalPValue), and q-value (qValue).

The log2AllelicFoldChange is the allelic fold change (in log2 scale) of the most significant sQTL.

The pValue is the nominal p-value of the most significant sQTL.

The pValueThreshold is the p-value threshold used to determine whether a cis-sQTL for this gene is significant. For more details see https://gtexportal.org/home/documentationPage#staticTextAnalysisMethods.

The empiricalPValue is the beta distribution-adjusted empirical p-value from FastQTL.

The qValues were calculated based on the empirical p-values. A false discovery rate (FDR) threshold of <= 0.05 was applied to identify genes with a significant sQTL.

See also

Examples

get_sqtl_genes("Whole_Blood")
#> Warning: ! Total number of items (3013) exceeds maximum page size (250).
#> ℹ Try increasing `itemsPerPage`.
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 13
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 3013
#> # A tibble: 250 × 11
#>    nPhenotypes pValueThreshold phenotypeId        geneSymbol    pValue datasetId
#>          <int>           <dbl> <chr>              <chr>          <dbl> <chr>    
#>  1           9      0.0000145  chr1:15947:16607:… WASH7P     7.15e-  8 gtex_v8  
#>  2           6      0.00000634 chr1:1090428:1091… C1orf159   6.85e- 10 gtex_v8  
#>  3           7      0.00000908 chr1:1228946:1231… SDF4       1.16e-188 gtex_v8  
#>  4           6      0.00000861 chr1:1310688:1310… PUSL1      2.50e- 40 gtex_v8  
#>  5           3      0.00000954 chr1:1325102:1327… CPTP       9.60e-  7 gtex_v8  
#>  6           2      0.0000147  chr1:1487914:1489… ATAD3C     1.61e- 25 gtex_v8  
#>  7           5      0.00000740 chr1:1487914:1489… ATAD3B     1.61e- 25 gtex_v8  
#>  8           2      0.0000361  chr1:1545002:1547… SSU72      3.28e- 25 gtex_v8  
#>  9          19      0.00000176 chr1:1629311:1629… MIB2       1.97e- 12 gtex_v8  
#> 10           8      0.00000273 chr1:1641113:1641… CDK11B     3.08e- 56 gtex_v8  
#> # ℹ 240 more rows
#> # ℹ 5 more variables: empiricalPValue <dbl>, tissueSiteDetailId <chr>,
#> #   ontologyId <chr>, qValue <dbl>, gencodeId <chr>