Get Transcripts

Find all transcripts of a reference gene.

This service returns information about transcripts of the given versioned GENCODE ID.

A genome build and GENCODE version must be provided.

By default, this service queries the genome build and GENCODE version used by the latest GTEx release.

GTEx API Portal documentation

Usage

get_transcripts(
  gencodeId,
  gencodeVersion = "v26",
  genomeBuild = "GRCh38/hg38",
  page = 0,
  itemsPerPage = 250
)

Arguments

gencodeId

String. A Versioned GENCODE ID of a gene, e.g. "ENSG00000065613.9".

gencodeVersion

String (default = "v26"). GENCODE annotation release. Either "v26" or "v19".

genomeBuild

String. Options: "GRCh38/hg38", "GRCh37/hg19". Default = "GRCh38/hg38".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A tibble.

See also

Examples

# \dontrun{
get_transcripts(gencodeId = "ENSG00000203782.5")
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 1
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 1
#> # A tibble: 1 × 11
#>    start    end featureType genomeBuild transcriptId source chromosome gencodeId
#>    <int>  <int> <chr>       <chr>       <chr>        <chr>  <chr>      <chr>    
#> 1 1.53e8 1.53e8 transcript  GRCh38/hg38 ENST0000036… HAVANA chr1       ENSG0000…
#> # ℹ 3 more variables: geneSymbol <chr>, gencodeVersion <chr>, strand <chr>
# }