Get Independent eQTL

Retrieve Independent eQTL Data

Finds and returns Independent eQTL Data data for the provided list of genes

By default, this endpoint fetches data from the latest GTEx version

The retrieved data is split into pages with items_per_page entries per page

GTEx portal API documentation

Usage

get_independent_eqtl(
  gencodeIds,
  tissueSiteDetailIds = NULL,
  datasetId = "gtex_v8",
  page = 0,
  itemsPerPage = 250
)

Arguments

gencodeIds

A character vector of Versioned GENCODE IDs, e.g. c("ENSG00000132693.12", "ENSG00000203782.5").

tissueSiteDetailIds

Character vector of IDs for tissues of interest. Can be GTEx specific IDs (e.g. "Whole_Blood"; use available_tissueSiteDetailIds() to see valid values) or Ontology IDs.

datasetId

String. Unique identifier of a dataset. Usually includes a data source and data release. Options: "gtex_v8", "gtex_snrnaseq_pilot". Default = "gtex_v8".

page

Integer (default = 0).

itemsPerPage

Integer (default = 250).

Value

A tibble.

See also

Examples

# search by gene
get_independent_eqtl(gencodeIds = c("ENSG00000132693.12",
                               "ENSG00000203782.5"))
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 1
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 8
#> # A tibble: 8 × 14
#>   gencodeId       geneSymbol variantId snpId tissueSiteDetailId ontologyId  rank
#>   <chr>           <chr>      <chr>     <chr> <chr>              <chr>      <int>
#> 1 ENSG0000013269… CRP        chr1_160… rs66… Esophagus_Gastroe… UBERON:00…     1
#> 2 ENSG0000013269… CRP        chr1_159… rs57… Thyroid            UBERON:00…     1
#> 3 ENSG0000020378… LOR        chr1_153… rs14… Artery_Coronary    UBERON:00…     1
#> 4 ENSG0000020378… LOR        chr1_153… rs87… Whole_Blood        UBERON:00…     1
#> 5 ENSG0000020378… LOR        chr1_153… rs12… Skin_Sun_Exposed_… UBERON:00…     1
#> 6 ENSG0000020378… LOR        chr1_152… rs77… Heart_Atrial_Appe… UBERON:00…     1
#> 7 ENSG0000020378… LOR        chr1_153… rs12… Esophagus_Mucosa   UBERON:00…     1
#> 8 ENSG0000013269… CRP        chr1_159… rs86… Muscle_Skeletal    UBERON:00…     1
#> # ℹ 7 more variables: tssDistance <int>, maf <dbl>, pValue <dbl>, nes <dbl>,
#> #   chromosome <chr>, pos <int>, datasetId <chr>

# optionally filter for a single variant and/or one or more tissues
get_independent_eqtl(gencodeIds = c("ENSG00000132693.12",
                                    "ENSG00000203782.5"),
                     tissueSiteDetailIds = c("Whole_Blood",
                                             "Thyroid"))
#> 
#> ── Paging info ─────────────────────────────────────────────────────────────────
#> • numberOfPages = 1
#> • page = 0
#> • maxItemsPerPage = 250
#> • totalNumberOfItems = 2
#> # A tibble: 2 × 14
#>   gencodeId       geneSymbol variantId snpId tissueSiteDetailId ontologyId  rank
#>   <chr>           <chr>      <chr>     <chr> <chr>              <chr>      <int>
#> 1 ENSG0000013269… CRP        chr1_159… rs57… Thyroid            UBERON:00…     1
#> 2 ENSG0000020378… LOR        chr1_153… rs87… Whole_Blood        UBERON:00…     1
#> # ℹ 7 more variables: tssDistance <int>, maf <dbl>, pValue <dbl>, nes <dbl>,
#> #   chromosome <chr>, pos <int>, datasetId <chr>